+++ packages/client/package.json
@@ -21,6 +21,7 @@
 	},
 	"devDependencies": {
 		"@react-three/postprocessing": "^2.15.11",
+		"@types/node": "^20.9.2",
 		"@types/react": "^18.2.15",
 		"@types/react-dom": "^18.2.7",
 		"@types/three": "^0.158.2",
