--- packages/client/src/constants/star.ts
@@ -1,1 +1,0 @@
-export const DIST_LIMIT = 3000;
