+++ packages/client/src/features/backgroundStars/lib/constants/index.ts
