+++ packages/client/src/features/star/lib/constants/index.ts
@@ -1,0 +1,1 @@
+export const DISTANCE_LIMIT = 3000;
