+++ packages/client/src/shared/lib/constants/camera.ts
