+++ packages/client/src/shared/lib/constants/index.ts
@@ -1,0 +1,2 @@
+export * from './camera';
+export * from './space';
