+++ packages/client/src/widgets/galaxy/lib/constants/index.ts
